requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
python-dotenv==1.0.0
gradio-client==0.8.1
aiohttp==3.9.1
Pillow==10.1.0
asyncio